Moved Metadata logik back to the page.tsx
